refracto + fix tests
