+++ test/client/client.basic.spec.ts
@@ -7,11 +7,11 @@ import {
   dropDatabases,
 } from '../src/app.module';
 import { CrudController } from '../../core/crud/crud.controller';
-import { MyUserService } from '../myuser.service';
+import { MyUserService } from '../src/services/myuser/myuser.service';
 import { CrudAuthService } from '../../core/authentication/auth.service';
 import { NestFastifyApplication } from '@nestjs/platform-fastify';
 import { EntityManager } from '@mikro-orm/mongodb';
-import { UserProfile } from '../entities/UserProfile';
+import UserProfile from '../src/services/userprofile/userprofile.entity';
 import {
   createAccountsAndProfiles,
   createMelons,
@@ -19,8 +19,8 @@ import {
   parseJwtCookieFromRes,
   testMethod,
 } from '../test.utils';
-import { MyProfileService } from '../profile.service';
-import { Melon } from '../entities/Melon';
+import { UserProfileService as MyProfileService } from '../src/services/userprofile/userprofile.service';
+import Melon from '../src/services/melon/melon.entity';
 import { TestUser } from '../test.utils';
 import {
   CRUD_CONFIG_KEY,
@@ -33,9 +33,9 @@ import {
   MemoryStorage,
 } from '../../client/CrudClient';
 import { LoginDto } from '../../core/crud/model/dtos';
-import { MelonService } from '../melon.service';
+import { MelonService } from '../src/services/melon/melon.service';
 import exp from 'constants';
-import { MyUser } from '../entities/MyUser';
+import MyUser from '../src/services/myuser/myuser.entity';
 
 const testAdminCreds = {
   email: 'admin@testmail.com',
