+++ test/client/client.cookie.spec.ts
@@ -7,11 +7,11 @@ import {
   dropDatabases,
 } from '../src/app.module';
 import { CrudController } from '@eicrud/core/crud/crud.controller';
-import { MyUserService } from '../myuser.service';
+import { MyUserService } from '../src/services/myuser/myuser.service';
 import { CrudAuthService } from '@eicrud/core/authentication/auth.service';
 import { NestFastifyApplication } from '@nestjs/platform-fastify';
 import { EntityManager } from '@mikro-orm/mongodb';
-import { UserProfile } from '../entities/UserProfile';
+import UserProfile from '../src/services/userprofile/userprofile.entity';
 import {
   createAccountsAndProfiles,
   createMelons,
@@ -20,8 +20,8 @@ import {
   parseJwtCookieFromRes,
   testMethod,
 } from '../test.utils';
-import { MyProfileService } from '../profile.service';
-import { Melon } from '../entities/Melon';
+import { UserProfileService as MyProfileService } from '../src/services/userprofile/userprofile.service';
+import Melon from '../src/services/melon/melon.entity';
 import { TestUser } from '../test.utils';
 import {
   CRUD_CONFIG_KEY,
@@ -34,9 +34,9 @@ import {
   MemoryStorage,
 } from '../../client/CrudClient';
 import { LoginDto } from '@eicrud/core/crud/model/dtos';
-import { MelonService } from '../melon.service';
+import { MelonService } from '../src/services/melon/melon.service';
 import exp from 'constants';
-import { MyUser } from '../entities/MyUser';
+import MyUser from '../src/services/myuser/myuser.entity';
 
 const testAdminCreds = {
   email: 'admin@testmail.com',
