+++ test/core/core.basic-nested.spec.ts
@@ -7,25 +7,25 @@ import {
   dropDatabases,
 } from '../src/app.module';
 import { CrudController } from '../../core/crud/crud.controller';
-import { MyUserService } from '../myuser.service';
+import { MyUserService } from '../src/services/myuser/myuser.service';
 import { CrudAuthService } from '../../core/authentication/auth.service';
 import { NestFastifyApplication } from '@nestjs/platform-fastify';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
-import { UserProfile } from '../entities/UserProfile';
+import UserProfile from '../src/services/userprofile/userprofile.entity';
 import {
   createAccountsAndProfiles,
   createNewProfileTest,
   testMethod,
 } from '../test.utils';
-import { MyProfileService } from '../profile.service';
+import { UserProfileService as MyProfileService } from '../src/services/userprofile/userprofile.service';
 import {
   CRUD_CONFIG_KEY,
   CrudConfigService,
 } from '../../core/config/crud.config.service';
 import { TestUser } from '../test.utils';
-import { Picture } from '../entities/Picture';
-import { Melon } from '../entities/Melon';
-import { MyPictureService } from '../picture.service';
+import Picture from '../src/services/picture/picture.entity';
+import Melon from '../src/services/melon/melon.entity';
+import { PictureService as MyPictureService } from '../src/services/picture/picture.service';
 import { CrudQuery } from '../../core/crud/model/CrudQuery';
 
 const testAdminCreds = {
