+++ test/core/core.cmd.spec.ts
@@ -7,14 +7,12 @@ import {
   dropDatabases,
 } from '../src/app.module';
 import { CrudController } from '../../core/crud/crud.controller';
-import { MyUserService } from '../myuser.service';
 import { CrudAuthService } from '../../core/authentication/auth.service';
 import {
   FastifyAdapter,
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
-import { UserProfile } from '../entities/UserProfile';
 import { CrudQuery } from '../../core/crud/model/CrudQuery';
 import {
   createAccountsAndProfiles,
@@ -22,8 +20,6 @@ import {
   createNewProfileTest,
   testMethod,
 } from '../test.utils';
-import { MyProfileService, TestCmdDto } from '../profile.service';
-import { Melon } from '../entities/Melon';
 import { CrudService } from '../../core/crud/crud.service';
 import { TestUser } from '../test.utils';
 import {
@@ -32,6 +28,11 @@ import {
 } from '../../core/config/crud.config.service';
 import { format } from 'path';
 import exp from 'constants';
+import { MyUserService } from '../src/services/myuser/myuser.service';
+import UserProfile from '../src/services/userprofile/userprofile.entity';
+import Melon from '../src/services/melon/melon.entity';
+import { UserProfileService as MyProfileService } from '../src/services/userprofile/userprofile.service';
+import TestCmdDto from '../src/services/userprofile/cmds/test_cmd/test_cmd.dto';
 
 const testAdminCreds = {
   email: 'admin@testmail.com',
