--- test/entities/FakeEmail.ts
@@ -1,23 +1,0 @@
-import { Property, Entity, PrimaryKey } from '@mikro-orm/core';
-import { CrudEntity } from '../../core/crud/model/CrudEntity';
-
-@Entity()
-export class FakeEmail implements CrudEntity {
-  @PrimaryKey({ name: '_id' })
-  id: string;
-
-  @Property()
-  to: string;
-
-  @Property()
-  message: string;
-
-  @Property()
-  type: string;
-
-  @Property()
-  createdAt: Date;
-
-  @Property()
-  updatedAt: Date;
-}
