+++ test/src/services/dragonfruit/cmds.ts
@@ -1,0 +1,3 @@
+//Auto generated file
+
+export const serviceCmds = {};
