+++ test/src/services/fakeemail/fakeemail.security.ts
@@ -1,0 +1,19 @@
+import { CrudSecurity } from '@eicrud/core/config';
+import { serviceCmds } from './cmds';
+
+export function getSecurity(fakeemail: string): CrudSecurity {
+  return {
+    rolesRights: {
+      guest: {
+        async defineCRUDAbility(can, cannot, ctx) {
+          // Define abilities for guest
+        },
+      },
+    },
+
+    cmdSecurityMap: Object.keys(serviceCmds).reduce((acc, cmd) => {
+      acc[cmd] = serviceCmds[cmd].getCmdSecurity(cmd, fakeemail);
+      return acc;
+    }, {}),
+  };
+}
