+++ test/src/services/melon/cmds/search/search.security.ts
@@ -1,0 +1,23 @@
+import { CmdSecurity } from '@eicrud/core/config';
+import search from './search.action';
+import SearchDto from './search.dto';
+
+const getCmdSecurity = (search, melon): CmdSecurity => {
+  return {
+    dto: SearchDto,
+    batchField: 'ids',
+    rolesRights: {
+      guest: {
+        maxBatchSize: 200,
+        async defineCMDAbility(can, cannot, ctx) {
+          can('search', melon);
+        },
+      },
+    },
+  };
+};
+
+export const searchSecurity = {
+  getCmdSecurity,
+  action: search,
+};
