--- test/src/services/melon/melon.service.spec.ts
@@ -1,40 +1,0 @@
-import { MongoDriver } from '@mikro-orm/mongodb';
-import { Test, TestingModule } from '@nestjs/testing';
-import { MelonService } from './melon.service';
-import { EICRUDModule } from '@eicrud/core';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
-import Melon from './melon.entity';
-import { MyConfigService } from '../../eicrud.config.service';
-import { CRUD_CONFIG_KEY } from '@eicrud/core/config';
-
-describe('AppController', () => {
-  let myService: MelonService;
-
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
-      imports: [
-        MikroOrmModule.forRoot({
-          entities: [Melon],
-          driver: MongoDriver,
-          dbName: 'test-melon',
-        }),
-        EICRUDModule.forRoot(),
-      ],
-      providers: [
-        MelonService,
-        {
-          provide: CRUD_CONFIG_KEY,
-          useClass: MyConfigService,
-        },
-      ],
-    }).compile();
-
-    myService = app.get<MelonService>(MelonService);
-  });
-
-  describe('root', () => {
-    it('should be defined"', () => {
-      expect(myService).toBeDefined();
-    });
-  });
-});
