+++ test/src/services/melon/melon.service.ts
@@ -1,3 +1,4 @@
+import SearchDto from './cmds/search/search.dto';
 import { ModuleRef } from '@nestjs/core';
 import Melon from './melon.entity';
 import { Injectable } from '@nestjs/common';
@@ -5,6 +6,7 @@ import { getSecurity } from './melon.security';
 import { CrudService } from '@eicrud/core/crud';
 import { serviceCmds } from './cmds';
 import { CrudContext } from '@eicrud/core/crud';
+import { CrudAuthorizationService } from '@eicrud/core/crud/crud.authorization.service';
 
 @Injectable()
 export class MelonService extends CrudService<Melon> {
@@ -14,4 +16,13 @@ export class MelonService extends CrudService<Melon> {
   }
 
   // GENERATED START - do not remove
+  async $search(dto: SearchDto, ctx: CrudContext, inheritance?: any) {
+    return await serviceCmds.search.action.call(
+      this,
+      dto,
+      this,
+      ctx,
+      inheritance,
+    );
+  }
 }
