+++ test/src/services/picture/picture.entity.ts
@@ -4,7 +4,7 @@ import { IsNumber, IsOptional, IsString, MaxLength } from 'class-validator';
 import UserProfile from '../userprofile/userprofile.entity';
 
 @Entity()
-export class Picture implements CrudEntity {
+export default class Picture implements CrudEntity {
   @PrimaryKey({ name: '_id' })
   @IsString()
   @IsOptional()
